fixing bugs when parse output in utility.py

+++ REST_FindSim/tasks/views.py
@@ -61,7 +61,7 @@ class OptimizationViewSet(viewsets.ModelViewSet):
         serializer.is_valid(raise_exception=True)
         serializer.save()
 
-        #parse the path to .tsv file and model file
+        # Parse the path to .tsv file and model file
         tsv_path = os.path.join(BASE_FILE_PATH, 'tsv/') + serializer.validated_data['tsv_files'].name.split('/').pop()
         model_path = os.path.join(BASE_FILE_PATH, 'model/') + serializer.validated_data['model_file'].name.split('/').pop()
         file_label = serializer.validated_data['username'] + str(time.time())
@@ -70,7 +70,8 @@ class OptimizationViewSet(viewsets.ModelViewSet):
         os.makedirs(os.path.join(BASE_FILE_PATH, 'model/')+file_label+'/optimized')
         tsv_path_new = os.path.join(BASE_FILE_PATH, 'tsv/'+file_label+'/') + serializer.validated_data['tsv_files'].name
         model_path_new = os.path.join(BASE_FILE_PATH, 'model/' + file_label+'/') + serializer.validated_data['model_file'].name
-        print(tsv_path_new)
+
+        print(serializer.validated_data['model_file'].__repr__())
         # Rename the files to avoid collision
         if os.path.exists(tsv_path):
             os.rename(tsv_path, tsv_path_new)
@@ -82,30 +83,33 @@ class OptimizationViewSet(viewsets.ModelViewSet):
         else:
             serializer.validated_data['error'] = 'model file not found'
 
-        # define optimized model name
-        optimized_model = os.path.join(BASE_FILE_PATH, 'model/') + file_label + '/optimized/' + model_path_new.split('/').pop()
-        optimized_model_url = os.path.join(BASE_FILE_PATH, 'model/') + file_label + '/optimized/'
-        optimized_model_url = optimized_model_url + serializer.validated_data['model_file'].name.split('.')[0] \
+        # If no errors, run optimization:
+        if 'error' not in serializer.validated_data:
+            # define optimized model name
+            optimized_model = os.path.join(BASE_FILE_PATH, 'model/') + file_label + '/optimized/' + model_path_new.split('/').pop()
+            optimized_model_url = os.path.join(BASE_FILE_PATH, 'model/') + file_label + '/optimized/'
+            optimized_model_url = optimized_model_url + serializer.validated_data['model_file'].name.split('.')[0] \
                               + '_tweaked.' + serializer.validated_data['model_file'].name.split('.')[1]
-        optimized_model_url = optimized_model_url[6:]
-        # Get num_processes & tolerance
-        tolerance = serializer.validated_data['tolerance']
-        num_processes = serializer.validated_data['num_processes']
-        # Run optimization via subprocess
-        res = run_optimization(tsv_path_new, model_path_new, file_label, optimized_model, num_processes, tolerance)
-        os.remove(model_path_new)
-        shutil.rmtree(os.path.join(BASE_FILE_PATH, 'tsv/')+file_label)
+            optimized_model_url = optimized_model_url[6:]
+            # Get num_processes & tolerance
+            tolerance = serializer.validated_data['tolerance']
+            num_processes = serializer.validated_data['num_processes']
+            # Run optimization via subprocess
+            res = run_optimization(tsv_path_new, model_path_new, file_label, optimized_model, num_processes, tolerance)
+            # Remove tsv files and model file
+            os.remove(model_path_new)
+            shutil.rmtree(os.path.join(BASE_FILE_PATH, 'tsv/')+file_label)
+            # Get result
+            serializer.validated_data['score'] = res.score
+            serializer.validated_data['time'] = res.time
+            serializer.validated_data['parameters'] = json.dumps(res.parameters)
+            # Get optimzed model
+            if not res.has_error():
+                res_model_file = open(res.model)
+                serializer.validated_data['optimized_model'] = optimized_model_url
+                res_model_file.close()
+            serializer.validated_data['error'] = res.error
 
-        serializer.validated_data['score'] = res.score
-        serializer.validated_data['time'] = res.time
-        serializer.validated_data['parameters'] = json.dumps(res.parameters)
-
-        if not res.has_error():
-            res_model_file = open(res.model)
-            serializer.validated_data['optimized_model'] = optimized_model_url
-            res_model_file.close()
-
-        serializer.validated_data['error'] = res.error
         serializer.save()
         headers = self.get_success_headers(serializer.data)
         os.remove(tsv_path)
